+++ data/types/settings/pump/bolus_calculator_insulin_test.go
@@ -44,12 +44,36 @@ var _ = Describe("BolusCalculatorInsulin", func() {
 		Expect(pump.BolusCalculatorInsulinDurationHoursMinimum).To(Equal(0.0))
 	})
 
+	It("BolusCalculatorInsulinDurationMinutesMaximum is expected", func() {
+		Expect(pump.BolusCalculatorInsulinDurationMinutesMaximum).To(Equal(600.0))
+	})
+
+	It("BolusCalculatorInsulinDurationMinutesMinimum is expected", func() {
+		Expect(pump.BolusCalculatorInsulinDurationMinutesMinimum).To(Equal(0.0))
+	})
+
+	It("BolusCalculatorInsulinDurationSecondsMaximum is expected", func() {
+		Expect(pump.BolusCalculatorInsulinDurationSecondsMaximum).To(Equal(36000.0))
+	})
+
+	It("BolusCalculatorInsulinDurationSecondsMinimum is expected", func() {
+		Expect(pump.BolusCalculatorInsulinDurationSecondsMinimum).To(Equal(0.0))
+	})
+
 	It("BolusCalculatorInsulinUnitsHours is expected", func() {
 		Expect(pump.BolusCalculatorInsulinUnitsHours).To(Equal("hours"))
 	})
 
+	It("BolusCalculatorInsulinUnitsMinutes is expected", func() {
+		Expect(pump.BolusCalculatorInsulinUnitsMinutes).To(Equal("minutes"))
+	})
+
+	It("BolusCalculatorInsulinUnitsSeconds is expected", func() {
+		Expect(pump.BolusCalculatorInsulinUnitsSeconds).To(Equal("seconds"))
+	})
+
 	It("BolusCalculatorInsulinUnits returns expected", func() {
-		Expect(pump.BolusCalculatorInsulinUnits()).To(Equal([]string{"hours"}))
+		Expect(pump.BolusCalculatorInsulinUnits()).To(Equal([]string{"hours", "minutes", "seconds"}))
 	})
 
 	Context("ParseBolusCalculatorInsulin", func() {
@@ -119,35 +143,35 @@ var _ = Describe("BolusCalculatorInsulin", func() {
 						datum.Units = pointer.FromString("invalid")
 					},
 					testErrors.WithPointerSource(structureValidator.ErrorValueNotExists(), "/duration"),
-					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours"}), "/units"),
+					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units invalid; duration out of range (lower)",
 					func(datum *pump.BolusCalculatorInsulin) {
 						datum.Duration = pointer.FromFloat64(-0.1)
 						datum.Units = pointer.FromString("invalid")
 					},
-					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours"}), "/units"),
+					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units invalid; duration in range (lower)",
 					func(datum *pump.BolusCalculatorInsulin) {
 						datum.Duration = pointer.FromFloat64(0.0)
 						datum.Units = pointer.FromString("invalid")
 					},
-					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours"}), "/units"),
+					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units invalid; duration in range (upper)",
 					func(datum *pump.BolusCalculatorInsulin) {
 						datum.Duration = pointer.FromFloat64(10.0)
 						datum.Units = pointer.FromString("invalid")
 					},
-					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours"}), "/units"),
+					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units invalid; duration out of range (upper)",
 					func(datum *pump.BolusCalculatorInsulin) {
 						datum.Duration = pointer.FromFloat64(10.1)
 						datum.Units = pointer.FromString("invalid")
 					},
-					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours"}), "/units"),
+					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units hours: duration missing",
 					func(datum *pump.BolusCalculatorInsulin) {
@@ -182,13 +206,79 @@ var _ = Describe("BolusCalculatorInsulin", func() {
 					},
 					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(10.1, 0.0, 10.0), "/duration"),
 				),
+				Entry("units minutes: duration missing",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = nil
+						datum.Units = pointer.FromString("minutes")
+					},
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotExists(), "/duration"),
+				),
+				Entry("units minutes: duration out of range (lower)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(-0.1)
+						datum.Units = pointer.FromString("minutes")
+					},
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(-0.1, 0.0, 600.0), "/duration"),
+				),
+				Entry("units minutes: duration in range (lower)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(0.0)
+						datum.Units = pointer.FromString("minutes")
+					},
+				),
+				Entry("units minutes: duration in range (upper)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(600.0)
+						datum.Units = pointer.FromString("minutes")
+					},
+				),
+				Entry("units minutes: duration out of range (upper)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(600.1)
+						datum.Units = pointer.FromString("minutes")
+					},
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(600.1, 0.0, 600.0), "/duration"),
+				),
+				Entry("units seconds: duration missing",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = nil
+						datum.Units = pointer.FromString("seconds")
+					},
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotExists(), "/duration"),
+				),
+				Entry("units seconds: duration out of range (lower)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(-0.1)
+						datum.Units = pointer.FromString("seconds")
+					},
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(-0.1, 0.0, 36000.0), "/duration"),
+				),
+				Entry("units seconds: duration in range (lower)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(0.0)
+						datum.Units = pointer.FromString("seconds")
+					},
+				),
+				Entry("units seconds: duration in range (upper)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(36000.0)
+						datum.Units = pointer.FromString("seconds")
+					},
+				),
+				Entry("units seconds: duration out of range (upper)",
+					func(datum *pump.BolusCalculatorInsulin) {
+						datum.Duration = pointer.FromFloat64(36000.1)
+						datum.Units = pointer.FromString("seconds")
+					},
+					testErrors.WithPointerSource(structureValidator.ErrorValueNotInRange(36000.1, 0.0, 36000.0), "/duration"),
+				),
 				Entry("units missing",
 					func(datum *pump.BolusCalculatorInsulin) { datum.Units = nil },
 					testErrors.WithPointerSource(structureValidator.ErrorValueNotExists(), "/units"),
 				),
 				Entry("units invalid",
 					func(datum *pump.BolusCalculatorInsulin) { datum.Units = pointer.FromString("invalid") },
-					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours"}), "/units"),
+					testErrors.WithPointerSource(structureValidator.ErrorValueStringNotOneOf("invalid", []string{"hours", "minutes", "seconds"}), "/units"),
 				),
 				Entry("units hours",
 					func(datum *pump.BolusCalculatorInsulin) {
@@ -248,10 +338,22 @@ var _ = Describe("BolusCalculatorInsulin", func() {
 			Expect(maximum).To(Equal(math.MaxFloat64))
 		})
 
-		It("returns expected range for units units", func() {
+		It("returns expected range for units hours", func() {
 			minimum, maximum := pump.BolusCalculatorInsulinDurationRangeForUnits(pointer.FromString("hours"))
 			Expect(minimum).To(Equal(0.0))
 			Expect(maximum).To(Equal(10.0))
+		})
+
+		It("returns expected range for units minutes", func() {
+			minimum, maximum := pump.BolusCalculatorInsulinDurationRangeForUnits(pointer.FromString("minutes"))
+			Expect(minimum).To(Equal(0.0))
+			Expect(maximum).To(Equal(600.0))
+		})
+
+		It("returns expected range for units seconds", func() {
+			minimum, maximum := pump.BolusCalculatorInsulinDurationRangeForUnits(pointer.FromString("seconds"))
+			Expect(minimum).To(Equal(0.0))
+			Expect(maximum).To(Equal(36000.0))
 		})
 	})
 })